requirements.txt
pyserial
pytest
python-dotenv
fastapi
uvicorn
